ai-service/requirements.txt
langgraph==0.2.0
langchain==0.2.0
langchain-openai==0.1.0
langchain-community==0.2.0
pika==1.3.2
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.5.0
redis==5.0.0
